ruuvibot/pyproject.toml
[tool.poetry]
name = "ruuvibot"
version = "0.5.6"
description = "Telegram bot to read ruuvitag data and log data to database"
authors = ["Rami Rahikkala <rami.rahikkala@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pexpect = "^4.8.0"
ruuvitag-sensor = "^2.0.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ruuvibot = 'ruuvibot:main'
